+++ packages/alloy/lib/src/registration/async_singleton_registration.dart
@@ -6,10 +6,12 @@ final class AsyncSingletonRegistration extends AlloyRegistration {
     required super.order,
     required this.factory,
     required this.dependsOn,
+    this.teardown,
   });
 
   final AlloyAsyncFactory<Object> factory;
   final Set<AlloyKey> dependsOn;
+  final AlloyTeardown? teardown;
 
   Object? instance;
   bool isReady = false;