+++ packages/alloy_analyzer/lib/src/parser/alloy_matchers.dart
@@ -7,3 +7,4 @@ const bootstrapMatcher = AlloyAnnotationMatcher('AlloyBootstrap');
 const initMatcher = AlloyAnnotationMatcher('AlloyInit');
 const environmentMatcher = AlloyAnnotationMatcher('AlloyEnvironment');
 const scopeRootMatcher = AlloyAnnotationMatcher('AlloyScopeRoot');
+const moduleMatcher = AlloyAnnotationMatcher('AlloyModule');