+++ packages/alloy_annotations/lib/alloy_annotations.dart
@@ -11,6 +11,7 @@ export 'package:alloy_annotations/src/alloy_environment.dart';
 export 'package:alloy_annotations/src/alloy_init.dart';
 export 'package:alloy_annotations/src/alloy_inject.dart';
 export 'package:alloy_annotations/src/alloy_lifetime.dart';
+export 'package:alloy_annotations/src/alloy_module.dart';
 export 'package:alloy_annotations/src/alloy_provided.dart';
 export 'package:alloy_annotations/src/alloy_scope_root.dart';
 export 'package:alloy_annotations/src/injected.dart';